net-im/gotktrix: add RESTRICT=mirror
gotktrix isn't in the main repos and there is a lot of additional
things that need to be downloaded. If someone has a lot of mirrors
in their make.conf this will severely slow down the install process

I'll be bringing this change to other ebuilds in this repo soon.